requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
Flask-Parameter-Validation==2.4.0
Flask-Script==2.0.6